# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/plans": {
      "post": {
        "operationId": "PlansController_create",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlansController_findAll",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/plans/{id}": {
      "get": {
        "operationId": "PlansController_findOne",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PlansController_update",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PlansController_remove",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/mobile-updates/check": {
      "get": {
        "operationId": "MobileUpdatesController_check",
        "summary": "Check if a mobile OTA update is available",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "android"
              ],
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "default": "stable",
              "enum": [
                "stable",
                "beta"
              ],
              "type": "string"
            }
          },
          {
            "name": "appVersion",
            "required": true,
            "in": "query",
            "schema": {
              "example": "1.0.0",
              "type": "string"
            }
          },
          {
            "name": "bundleVersion",
            "required": true,
            "in": "query",
            "schema": {
              "example": "web-20260213-001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ]
      }
    },
    "/v1/mobile-updates/events": {
      "post": {
        "operationId": "MobileUpdatesController_logEvent",
        "summary": "Log mobile OTA update telemetry event",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMobileUpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ]
      }
    },
    "/v1/mobile-updates/releases": {
      "get": {
        "operationId": "MobileUpdatesController_listReleases",
        "summary": "List mobile OTA releases",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "android"
              ],
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "stable",
                "beta"
              ],
              "type": "string"
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Filter by publication state",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MobileUpdatesController_createRelease",
        "summary": "Create a mobile OTA release metadata record",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMobileReleaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/mobile-updates/releases/{id}/upload": {
      "post": {
        "operationId": "MobileUpdatesController_uploadBundle",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "bundle": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "bundle"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/mobile-updates/releases/{id}/publish": {
      "patch": {
        "operationId": "MobileUpdatesController_publishRelease",
        "summary": "Publish or unpublish mobile OTA release",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishMobileReleaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/mobile-updates/releases/{id}/bundle": {
      "get": {
        "operationId": "MobileUpdatesController_downloadBundle",
        "summary": "Download published OTA bundle",
        "parameters": [
          {
            "name": "X-Client-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mobile Updates"
        ]
      }
    }
  },
  "info": {
    "title": "SubMan API",
    "description": "Reusable multi-tenant subscription API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "499.00"
          },
          "currency": {
            "type": "string",
            "default": "PHP"
          },
          "interval": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "YEARLY"
            ]
          },
          "intervalCount": {
            "type": "number",
            "default": 1
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "price",
          "interval"
        ]
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "499.00"
          },
          "currency": {
            "type": "string",
            "default": "PHP"
          },
          "interval": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "YEARLY"
            ]
          },
          "intervalCount": {
            "type": "number",
            "default": 1
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "CreateMobileUpdateEventDto": {
        "type": "object",
        "properties": {
          "releaseId": {
            "type": "string"
          },
          "eventType": {
            "type": "string",
            "enum": [
              "CHECKED",
              "DOWNLOADED",
              "APPLIED",
              "FAILED"
            ]
          },
          "appVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "bundleVersion": {
            "type": "string",
            "example": "web-20260213-001"
          },
          "deviceId": {
            "type": "string",
            "example": "android-12345"
          },
          "message": {
            "type": "string",
            "example": "Checksum mismatch during apply."
          }
        },
        "required": [
          "eventType",
          "appVersion",
          "bundleVersion"
        ]
      },
      "CreateMobileReleaseDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "android"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "stable",
              "beta"
            ],
            "default": "stable"
          },
          "appVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "bundleVersion": {
            "type": "string",
            "example": "web-20260213-001"
          },
          "releaseNotes": {
            "type": "string",
            "example": "Minor UI improvements and bug fixes."
          },
          "mandatory": {
            "type": "boolean",
            "default": false
          },
          "minimumSupportedAppVersion": {
            "type": "string",
            "example": "1.0.0"
          }
        },
        "required": [
          "platform",
          "appVersion",
          "bundleVersion"
        ]
      },
      "PublishMobileReleaseDto": {
        "type": "object",
        "properties": {
          "publish": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "publish"
        ]
      }
    }
  }
}